# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dictee
    description: Opérations de dictée
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dictee/ocr:
    post:
      operationId: extractWords
      tags: [dictee]
      summary: Extraire les mots d'une image
      description: Analyse une image et retourne la liste des mots détectés (OCR)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OcrRequest"
      responses:
        "200":
          description: Mots extraits avec succès
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OcrResponse"
        "400":
          description: Requête invalide
        "500":
          description: Erreur lors de l'analyse
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    OcrRequest:
      type: object
      properties:
        image:
          type: string
          description: Image encodée en base64
        mimeType:
          type: string
          description: Type MIME de l'image
          default: "image/jpeg"
      required:
        - image
    OcrResponse:
      type: object
      properties:
        mots:
          type: array
          items:
            type: string
          description: Liste des mots détectés
      required:
        - mots