# lib/api-zod/src/generated/types/ocrRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface OcrRequest {
  /** Image encodée en base64 */
  image: string;
  /** Type MIME de l'image */
  mimeType?: string;
}